Be more careful about trailing_slash in Gutenberg's get_url
If the user visits a directory without the trailing slash on its name, Apache
will redirect them to the directory *with* the trailing slash, as per:

http://httpd.apache.org/docs/2.2/mod/mod_dir.html

But when that happens, the protocol changes from HTTPS to HTTP, I think
because of newton.cx's haproxy setup where the communication between the proxy
and Apache doesn't go over SSL. That's undesirable, so take care to use
trailing slashes when needed.

The Gutenberg documentation seems to be a bit misleading; it implies that the
default is trailing_slash=true, while it seems to be false. Perhaps it was
different in the past.