Fix all the broken links.
I set `check_external_links = true` in `config.toml` and got everything
working. It takes about 1 minute to do all the checking.

I added a shortcode called `ulink`, for "unmanaged link" — it's used for links
that are internal to the site, but not pointing to content managed by
Gutenberg. It doesn't know how to check those links as-is but I don't want to
make them all absolute.